# docs/swagger.yaml
basePath: /
definitions:
...
  models.AuthProvider:
    enum:
    - local
    - google
    - github
    type: string
    x-enum-varnames:
    - LocalAuth
    - GoogleAuth
    - GitHubAuth
...
  models.UserCreateInput:
    properties:
      authProvider:
        $ref: '#/definitions/models.AuthProvider'
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      organizationIds:
        items:
          type: string
        type: array
      password:
        minLength: 8
        type: string
      providerUserId:
        type: string
      roleIds:
        items:
          type: string
        type: array
      username:
        type: string
    required:
    - authProvider
    - email
    - firstName
    - lastName
    - username
    type: object
  models.UserLoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.UserResponse:
    properties:
      active:
        type: boolean
      authProvider:
        $ref: '#/definitions/models.AuthProvider'
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      firstName:
        type: string
      id:
        type: string
      lastLogin:
        type: string
      lastName:
        type: string
      organizationIds:
        items:
          type: string
        type: array
      roleIds:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
  models.UserSwaggerResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: This is the API documentation for the RBAC system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: RBAC System API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user account with the provided details
      parameters:
      - description: User Login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginInput'
      produces:
      - application/json
      responses:
        "201":
          description: User Logged successfully
          schema:
            $ref: '#/definitions/models.UserSwaggerResponse'
        "400":
          description: Invalid request body or missing fields
        "409":
          description: Username or email already exists
        "500":
          description: Internal server error
      summary: Login a new user
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes an access token using the provided valid token
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/models.UserSwaggerResponse'
        "400":
          description: Invalid request or missing token
        "401":
          description: Unauthorized or invalid token
        "403":
          description: User account is disabled
        "500":
          description: Internal server error
      summary: Refresh access token
      tags:
      - auth
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a valid refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: refreshToken
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated successfully
        "400":
          description: Invalid request body or missing refresh token
        "401":
          description: Invalid or expired refresh token
        "500":
          description: Internal server error
      summary: Refresh Access Token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided details
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/models.UserSwaggerResponse'
        "400":
          description: Invalid request body or missing fields
        "409":
          description: Username or email already exists
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"